골-멤버 조회 시 서비스 레이어에 트랜젝션 묶음
